Sidetables should not use `operator new` or `operator delete`.
